openstack/CleanWlmDatabase/wlSnapshotDeletePython/lib/requirement.txt
mysql-connector-python==8.0.21
